Work in many areas.
Added PostOperation; which will eventually be a way to save POST operations to CoreData for later transmission.
Added a subhead to the LoginHeaderCell, which helps inform the user as to why they need to log in.
Added timestamps to tweets. Still no timer.
Added ComposeTweetController, and some of its cells.
Added Settings controller, and some of its cells.
DataSources are getting closer to where I want them.
Some basic Reachability-type stuff is in, but needs more work.