FetchedResultsController now uses cellmodels.
This was done mostly because the FRC datasource needed to cache cell sizes so it doesn’t recompute them all the time, and CellModels already support storing cell sizes, and we’re already got a lot of infrastructure for cell models.

Not done yet, as we need to add code to update the cell model array in response to FRC updates.