Custom Camera.
Added custom camera VC, with a bunch of new png images.

Changes to the photo selection cell to support it, as well as making the PhotoLibrary datasource a KrakenDataSourceSegment so that we can (easily) combine photo cells with cells from other sources in a single DS.

Added a refresh timer and notification for updating relative time values where they’re displayed. Twitarr and Seamail cells now use this.

Refactored how reactions work, again. Changes are easier to observe now.

ImageManager got a new fn to resize/reformat images in prep for upload.