Schedule Improvements.
Follow, Local Notification, and Set Calendar Events for Events all work.
Scroll to Now works, as does scroll to day of week.

Still need to write custom notification code.

Filtering is getting better; can now filter on events with calendar/alarmclock/follow state.

Added some (janky) global nav code so tapping on a notification can nav to the item in the app that caused the notification. In this case, tapping on an event reminder navs to the event.