Added timed server update calls.
We now contact the server to check server time and disabled app sections on a regular basis. Alerts should get added to this system. Also going to do a thing where the active view gets to run updates on a timer as well.

Added DisabledContentViewController, which dynamically replaces tabs in the tab bar when sections of the app are server-disabled.

The Time section of Settings now has real-ish information in it.