Updated UICollectionViewCells
Something in iOS 13 makes it so collection views try to size all their cells upon creation, when their size is 0, 0. So, all the collectionView cells now have autolayout constraints set so that their layout solves for 0-sized cells.

Generally, the fix was to set bottom and trailing constraints that go to the cell bounds to a 900 priority. Note that this might cause weirdness if we get cells with really wacky sizes (like 10, 10). But that should not happen, and even in the 0,0 case we’re seeing, there aren’t any cells that appear with that size; it’s just a side effect of collectionview creation.

Also: Whatever is causing this to happen during cv creation may itself be a bug; dunno.