Forums are coming along.
Also, modified how Segues work—we now wrap performSegue in the base VC, and we have an enum with known segues. Individual VCs need to declare what segues they support. Sort of a hack way to do it since you now need to add a value to knownSegues whenever you add a segue, but this gets us the ability to query what segues are supported from any VC.

The tweetCell got reworked. Instead of binding on the model we now bind a bunch of individual values that are much closer to the UI widgets. This is going to make it easier to build cell models for forum posts and forum postOps.

Forums got their filter types added. Need to do a bit more in FRCDataSource to support it.