Moved most segues to global segue controller.
Most segues in the app setup the destination VC soley from data in the sender parameter—they don’t need the source VC at all. Yet, the source VC is the only place where we can put the prepare:forSegue: method.

So, the logic that sets up destination data for segues is now all in the base VC. Eventually I may change this so that there’s a method in the dest VC that takes the sender param, and the root VC can call it with the sender.

Also, the forum compose VC is halfway built, and deep sea mode is coming along.