Forums has cache management.
We remember what forums we’ve loaded recently, and what posts when looking at a thread.

We could add code so we’d remember the loaded post counts so we don’t reload posts when the user leaves and re-enters a thread in a short time period.

Also, you can upload one pic in a forum post. Still need multi-pic.