Refactored TwitarrDataManager.
Finally added the FilterPack class, which combines a NSPredicate for getting filtered results from CoreData with a query URL generator that can ask the server for tweets based on the same query.

The caching timeout code has moved into this class as well. This way, each view on the Twittar data make one of these objects and uses it to source tweets. The object then makes server calls as necessary and caches when it made them and what tweets are ‘fresh’.

Seamail badging is in. I had to add calls to /api/v2/seamail/:thread_id to do it, as otherwise seamail threads aren’t ever marked read on the server.

Pursurant to that, we now call the Alerts endpoint.