Modified global nav. Moved to 5 tab system.
Now globalNav starts with the app delegate.

The app no longer has a More tab. Also, iPad versions can work in a way such that DailyVC is the root, and all top-level app features are nav pushes from that.

I haven’t gotten to iPad support yet, but the idea is that it’ll be a container VC with ~3 sub VC columns. Each column should be a nav stack rooted with a DailyVC.