First pass at Accessibility customizations.
Mostly this is cases where we should speak the contents of multiple labels in succession.

Also added cases where an overlay view should block accessibility nav to the underlying collection view while it’s open (such as the Filter view in Events).

Updated the song catalog, plus a few bugfixes.