Fixed an issue with PostOp delivery.
When you’re ashore, PostOps are supposed to just pile up, and get delivered when the network comes back, but when the network comes back they should be delivered in order, somewhat slowly. The code that handles this was hit by a recursion issue, leading to all the stored ops being delivered at once—instead of the intended 1 op/second.

Also changed CurrentUser.lastError to a Error, so that network errors can be delivered by this mechanism.