Core support for multiple pics in tweets.
Also, support in both forum posts and tweets for a mix of local image attachments that need to be uploaded when posting, and image files already on the server. This lets a user, for example, edit a tweet with an image, add a second image to it, and only the second image gets uploaded when the edit is posted. Mostly this was done by extending PhotoDataType to include more enum cases.

Still no UI for attaching multiple images to posts and twarrts, but this gets us closer.

Also, Network Governor should now log a user out locally when it receives a 401 response for any call.