Made viewing Twitarr replies better.
Doing so required that we change how Replying to a tweet works, in the case where there's already replies. The first to reply can do so right from the tweet cell; if the tweet has replies already the user instead must tap "Thread" to view the thread of replies, and can then tap "Reply" in the nav bar to reply.