Fixes rotation issue with uploaded videos
Portrait videos taken from an iPhone are stored landscape with a rotation matrix attached specifying that they should be rotated for display. AVComposition by default does not honor this when adding media to a composition (I consider this a bug since they honor *everything else* about the source media). Turns out you have to make a VideoCompositor and set the affine transform for each clip, and then add the compositor to the AVPlayerItem or the AssetExportSession. Anyway, it works now.

Also added code to killswitch MicroKaraoke from the server.