2024 Deck and Karaoke Updates
HAL has a new deck PDF for the Nieuw Amsterdam; the update changes a few room names. So, new deck plan PDFs, which means changes to DeckDataManager and LFGs, because of the named rooms.

Updated the Karaoke list for 2024; new list has ~300 new songs. For some reason, this year's list doesn't have the MIDI or Voice Reduced flags.

Updated the about text.

iOS 17 changed the rules for how calendar access works. So, new keys in the info.plist.