2025 Games list, networking fixes
Some no-auth-required server calls produce HTTP 401 when the server is in admin-only mode. This was causing Kraken to invalidate the current login. Now we only invalidate the login if we tried to auth and failed.

The day planner section header bars were using the wrong fn to get their dates.